refactor authentication context and API integration
