+++ src/types/auth/AuthContextType.ts
@@ -1,0 +1,6 @@
+export interface AuthContextType {
+    isAuthenticated: boolean;
+    setIsAuthenticated: (value: boolean) => void;
+    checkIsAuthenticated: () => boolean;
+    logout: () => void;
+}
