+++ src/types/auth/AuthResponse.ts
@@ -1,0 +1,3 @@
+export interface AuthResponse {
+    accessToken: string;
+}
