+++ src/types/auth/RegisterCredentials.ts
@@ -1,0 +1,6 @@
+export interface RegisterCredentials {
+    email: string;
+    password: string;
+    firstName?: string;
+    lastName?: string;
+}
