+++ src/types/user/UserProfile.ts
@@ -1,0 +1,6 @@
+export interface UserProfile {
+    id?: string;
+    email?: string;
+    firstName?: string;
+    lastName?: string;
+}
